Revert "[FEAT] 관리자 정보 회원가입 API개선"
This reverts commit 5826b3b360429e6845f325c6a4d40b703e36f662.
