--- src/main/java/com/backend/nova/admin/dto/AdminRefreshTokenRequest.java
@@ -1,6 +1,0 @@
-package com.backend.nova.admin.dto;
-
-public record AdminRefreshTokenRequest(
-        String refreshToken
-) {
-}
