--- src/main/java/com/backend/nova/admin/dto/AdminTokenResponse.java
@@ -1,12 +1,0 @@
-package com.backend.nova.admin.dto;
-
-import lombok.Builder;
-
-@Builder
-public record AdminTokenResponse(
-        String accessToken,
-        String refreshToken,
-        Long adminId,
-        String name
-) {
-}
